requirements.txt
click>=8.0.0
opencv-python>=4.5.0
paddleocr>=2.6.0
paddlepaddle>=2.4.0
python-pptx>=0.6.21
numpy>=1.21.0
Pillow>=9.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pymupdf>=1.21.0
scikit-image>=0.19.0
